+++ Slider/Assets/Materials/MagitechGemGoo.mat
@@ -58,7 +58,7 @@ Material:
         m_Offset: {x: 0, y: 0}
     m_Ints: []
     m_Floats:
-    - _ScrollSpeed: 0.1
+    - _ScrollSpeed: -0.1
     m_Colors:
     - _ColorMask: {r: 0.1019608, g: 0.1019608, b: 0.1019608, a: 0}
   m_BuildTextureStacks: []
