+++ Slider/Assets/Scripts/Map/MagiTech/GemManager.cs
@@ -16,6 +16,7 @@ public class GemManager : MonoBehaviour, ISavable
     public bool HasGemTransporter => hasGemTransporter;
     
     public PipeLiquid pipeLiquid;
+    public Animator animator;
 
     public void Save()
     {
@@ -123,10 +124,6 @@ public class GemManager : MonoBehaviour, ISavable
             Debug.LogWarning("Tried to turn in invalid item: " + item);
         }
         UpdateGemSprites();
-        // if(HasAllGems())
-        // {
-        //     EnableGFuel();
-        // }
     }
 
     public void UpdateGemSprites()
@@ -135,10 +132,15 @@ public class GemManager : MonoBehaviour, ISavable
         {
             gemSprites[a].SetActive(gems[a]);
         }
+        if(HasAllGems())
+        {
+            animator.Play("Active");
+        }
     }
 
     private bool HasAllGems()
     {
+        return true;
         foreach (bool b in gems.Values)
         {
             if (!b)
@@ -151,6 +153,7 @@ public class GemManager : MonoBehaviour, ISavable
 
     public void EnableGFuel(bool fillImmediate = false)
     {
+        if(pipeLiquid.isFilling || pipeLiquid.isFull) return;
         if(fillImmediate)
             pipeLiquid.SetPipeFull();
         else
