+++ Slider/Assets/Scripts/UI/Artifact/AreaArtifacts/MagiTech.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c56bc6162890a5428ab5586da5ad3ff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
