+++ Slider/Assets/Scripts/UI/Artifact/AreaArtifacts/MagiTech/MagitechUITrackerManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ca199b59e0ca96143b7c2deef957c185
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
