--- Slider/Assets/Scripts/UI/Artifact/AreaArtifacts/MagiTechArtifact.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8ebe6f78de9b35544887583d87f0fae4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
